refactor: パッケージ整理のため、リポジトリ検索関連の UI を ui.search パッケージに移動した。
