+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/search/RepositorySearchViewModel.kt
@@ -1,7 +1,4 @@
-/*
- * Copyright © 2021 YUMEMI Inc. All rights reserved.
- */
-package jp.co.yumemi.android.code_check
+package jp.co.yumemi.android.code_check.ui.search
 
 import android.content.Context
 import androidx.lifecycle.ViewModel
@@ -12,7 +9,8 @@ import io.ktor.client.request.get
 import io.ktor.client.request.header
 import io.ktor.client.request.parameter
 import io.ktor.client.statement.HttpResponse
-import jp.co.yumemi.android.code_check.TopActivity.Companion.lastSearchDate
+import jp.co.yumemi.android.code_check.R
+import jp.co.yumemi.android.code_check.TopActivity
 import jp.co.yumemi.android.code_check.data.model.RepositoryItem
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.async
@@ -68,7 +66,7 @@ class RepositorySearchViewModel(
                 )
             }
 
-            lastSearchDate = Date()
+            TopActivity.lastSearchDate = Date()
 
             return@async items.toList()
         }.await()
