+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/repositorySearchFragment"
-        android:name="jp.co.yumemi.android.code_check.RepositorySearchFragment"
+        android:name="jp.co.yumemi.android.code_check.ui.search.RepositorySearchFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_repository_search">
         <action
